+++ packages/fluentui/docs/src/routes.tsx
@@ -101,6 +101,11 @@ const Routes = () => (
             <Route exact path="/virtualized-tree" component={VirtualizedTreePrototype} />
             <Route exact path="/virtualized-table" component={VirtualizedTablePrototype} />
             <Route exact path="/prototype-copy-to-clipboard" component={CopyToClipboardPrototype} />
+            <Route
+              exact
+              path="/unstable-datepicker"
+              render={() => <Redirect to={`/components/datepicker/definition`} />}
+            />
             <Route exact path="/faq" component={FAQ} />
             <Route exact path="/accessibility" component={Accessibility} />
             <Route exact path="/accessibility-behaviors" component={AccessibilityBehaviors} />